incorporate Languages Nav into Languages page
