+++ .eslintrc.yml
@@ -81,6 +81,7 @@ rules:
     - warn
     - code: 120
       ignoreComments: true
+      ignoreStrings: true
   max-nested-callbacks:
     - warn
     - 3
@@ -91,7 +92,7 @@ rules:
     - max: 2
   new-cap: warn
   newline-per-chained-call: warn
-  no-alert: warn
+  no-alert: off
   no-array-constructor: warn
   no-bitwise: warn
   no-caller: error
