+++ src/pages/Languages/Languages.hbs
@@ -1,6 +1,10 @@
 <main id=main class='main languages-page' data-page=Languages>
 
-  {{> LanguagesNav }}
+  <nav class=languages__nav>
+    <h1 class=block-header>Languages</h1>
+    <ul class='js-languages-page__languages-list languages-page__languages-list'></ul>
+    <button class='button green js-languages-page__nav-add-lang-button languages-page__nav-add-lang-button'>Add a language</button>
+  </nav>
 
   <section class=language-editor></section>
