--- src/pages/Languages/LanguagesNav/LanguagesNav.hbs
@@ -1,5 +1,0 @@
-<nav class=languages-nav>
-  <h1 class=block-header>Languages</h1>
-  <ul class=languages-list></ul>
-  <button class='button green js-languages-nav__add-language-button languages-nav__add-language-button' type=button>Add a Language</button>
-</nav>
