--- src/pages/Languages/LanguagesNav/LanguagesNav.less
@@ -1,22 +1,0 @@
-@import (reference) 'utilities.less';
-@import (reference) 'variables.less';
-
-.languages-nav {
-
-  align-items: stretch;
-
-  &__add-language-button {
-    .align-center;
-    margin    : var(--space);
-  }
-
-  .list {
-    margin-inline: var(--space);
-  }
-
-  // large screens
-  @media (min-width: @breakpoint) {
-    width: 20em;
-  }
-
-}
